pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "axon"
version = "0.1.0"
description = "CLI image generator with terminal rendering"
requires-python = ">=3.9"
dependencies = ["Pillow>=10.0"]

[project.scripts]
axon = "axon.cli:main"
